refactor: rename legacy NIM test and mock files
Rename modelServingNim.cy.ts to legacyModelServingNim.cy.ts and
mockNimResource.ts to mockLegacyNimResource.ts to signal these
are legacy NIM code that will be replaced. Update all import paths.
