+++ frontend/src/__mocks__/mockLegacyNimResource.ts
