+++ packages/cypress/cypress/support/commands/odh.ts
@@ -86,7 +86,7 @@ import type {
   PipelineVersionKF,
 } from '@odh-dashboard/internal/concepts/pipelines/kfTypes';
 import type { GrpcResponse } from '@odh-dashboard/internal/__mocks__/mlmd/utils';
-import type { NimServingResponse } from '@odh-dashboard/internal/__mocks__/mockNimResource';
+import type { NimServingResponse } from '@odh-dashboard/internal/__mocks__/mockLegacyNimResource';
 import type { BuildMockPipelinveVersionsType } from '@odh-dashboard/internal/__mocks__';
 import type { ArtifactStorage } from '@odh-dashboard/internal/concepts/pipelines/types';
 import type { ConnectionTypeConfigMap } from '@odh-dashboard/internal/concepts/connectionTypes/types';
