+++ packages/cypress/cypress/tests/mocked/projects/tabs/legacyModelServingNim.cy.ts
@@ -7,7 +7,7 @@ import {
   mockMultipleNimServingRuntimes,
   mockNimImages,
   mockNimServingResource,
-} from '@odh-dashboard/internal/__mocks__/mockNimResource';
+} from '@odh-dashboard/internal/__mocks__/mockLegacyNimResource';
 import { InferenceServiceModel, ServingRuntimeModel } from '../../../../utils/models';
 import { projectDetails, projectDetailsOverviewTab } from '../../../../pages/projects';
 import { nimDeployModal } from '../../../../pages/components/NIMDeployModal';
